Added JavaDoc comments to the code
