+++ src/main/java/Duke/EmptyToDo.java
@@ -1,4 +1,7 @@
 package Duke;
 
+/**
+ * Error if description of Todo task is not given in user input
+ */
 public class EmptyToDo extends Exception{
 }
