+++ src/main/java/Duke/IllegalCommand.java
@@ -1,4 +1,7 @@
 package Duke;
 
+/**
+ * Error if user command is not recognised
+ */
 public class IllegalCommand extends Exception{
 }
